Удалить В MovieQuizPresenter конструкцию: `viewController?.alertModel = alertPresenter.createAlert(correct: correctAnswers, total: questionsAmount, message: text)`
