+++ MovieQuizPresenterUITests/MovieQuizPresenterUITests.swift
@@ -1,0 +1,80 @@
+//
+//  MovieQuizPresenterUITests.swift
+//  MovieQuizPresenterUITests
+//
+//  Created by Gleb on 19.02.2024.
+//
+
+import Foundation
+import XCTest
+@testable import MovieQuiz
+
+
+final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
+    var yesButton: UIButton!
+    
+    var noButton: UIButton!
+    
+    var alertModel: MovieQuiz.AlertModel?
+    
+    var activityIndicator: UIActivityIndicatorView!
+    
+    func show(quiz step: MovieQuiz.QuizStepViewModel) {
+    }
+    
+    func showAlert(quiz result: MovieQuiz.AlertModel) {
+    }
+    
+    func highlightImageBorder(isCorrectAnswer: Bool) {
+    }
+    
+    func showNetworkError(message: String) {
+    }
+}
+
+
+final class MovieQuizPresenterUITests: XCTestCase {
+    
+    override func setUpWithError() throws {
+        // Put setup code here. This method is called before the invocation of each test method in the class.
+        
+        // In UI tests it is usually best to stop immediately when a failure occurs.
+        continueAfterFailure = false
+        
+        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
+    }
+    
+    override func tearDownWithError() throws {
+        // Put teardown code here. This method is called after the invocation of each test method in the class.
+    }
+    
+    func testExample() throws {
+        // UI tests must launch the application that they test.
+        let app = XCUIApplication()
+        app.launch()
+        
+        // Use XCTAssert and related functions to verify your tests produce the correct results.
+    }
+    
+    func testLaunchPerformance() throws {
+        if #available(macOS 10.15, iOS 13.0, tvOS 13.0, watchOS 7.0, *) {
+            // This measures how long it takes to launch your application.
+            measure(metrics: [XCTApplicationLaunchMetric()]) {
+                XCUIApplication().launch()
+            }
+        }
+    }
+    
+    func testPresenterConvertModel() throws {
+        let viewControllerMock = MovieQuizViewControllerMock()
+        let sut = MovieQuizPresenter(viewController: viewControllerMock)
+        
+        let emptyData = Data()
+        let question = QuizQuestion(image: emptyData, text: "Question Text", correctAnswer: true)
+        let viewModel = sut.convert(model: question)
+        
+        XCTAssertNotNil(viewModel.image)
+        XCTAssertEqual(viewModel.question, "Question Text")
+        XCTAssertEqual(viewModel.questionNumber, "1/10")
+    }
+}
